LayoutTests/imported/w3c/web-platform-tests/css/css-values/getComputedStyle-border-radius-003-expected.txt


FAIL testing border-radius shorthand assert_equals: expected "calc(1% + 1px) calc(2% + 2px) calc(3% + 3px) calc(4% + 4px) / calc(5% + 5px) calc(6% + 6px) calc(7% + 7px) calc(8% + 8px)" but got "1px 2px 3px 4px / 5px 6px 7px 8px"
